fix: create a wrapper for google cal fetching
