+++ packages/app-store/googlecalendar/api/callback.ts
@@ -15,6 +15,7 @@ import getInstalledAppPath from "../../_utils/getInstalledAppPath";
 import { decodeOAuthState } from "../../_utils/oauth/decodeOAuthState";
 import { REQUIRED_SCOPES, SCOPE_USERINFO_PROFILE } from "../lib/constants";
 import { getGoogleAppKeys } from "../lib/getGoogleAppKeys";
+import { getAllCalendars } from "../lib/utils";
 
 async function getHandler(req: NextApiRequest, res: NextApiResponse) {
   const { code } = req.query;
@@ -74,11 +75,7 @@ async function getHandler(req: NextApiRequest, res: NextApiResponse) {
 
     const cals = await getAllCalendars(calendar);
 
-    let primaryCal = cals.find((cal) => cal.primary);
-    if (!primaryCal?.id) {
-      // If the primary calendar is not set, set it to the first calendar
-      primaryCal = cals[0];
-    }
+    const primaryCal = cals.find((cal) => cal.primary) ?? cals[0];
 
     // Only attempt to update the user's profile photo if the user has granted the required scope
     if (grantedScopes.includes(SCOPE_USERINFO_PROFILE)) {
@@ -202,24 +199,6 @@ async function updateProfilePhoto(oAuth2Client: Auth.OAuth2Client, userId: numbe
   } catch (error) {
     logger.error("Error updating avatarUrl from google calendar connect", error);
   }
-}
-
-async function getAllCalendars(calendar: calendar_v3.Calendar) {
-  let allCalendars: calendar_v3.Schema$CalendarListEntry[] = [];
-  let pageToken = null;
-
-  do {
-    const response: any = await calendar.calendarList.list({
-      fields: "items(id,summary,primary,accessRole),nextPageToken",
-      pageToken: pageToken,
-      maxResults: 250, // 250 is max
-    });
-
-    allCalendars = [...allCalendars, ...(response.data.items ?? [])];
-    pageToken = response.data.nextPageToken;
-  } while (pageToken);
-
-  return allCalendars;
 }
 
 export default defaultHandler({
